// src/pages/DandiBrowser/types.ts
export type AssetResponse = {
    access: {
        status: string
        schemaKey: string
    }[]
    approach: {
        name: string
        schemaKey: string
    }[]
    keywords: string[]
    schemaKey: string
    dateModified: string
    schemaVersion: string
    encodingFormat: string
    wasGeneratedBy: {
        name: string
        schemaKey: string
        startDate: string
        identifier: string
        description: string
    }[]
    wasAttributedTo: any // Co-pilot AI didn't help me with this one because there was a nested "sex" field
    blobDateModified: string
    variableMeasured: {
        value: string
        schemaKey: string
    }[]
    measurementTechnique: {
        name: string
        schemaKey: string
    }[]
    id: string
    path: string
    identifier: string
    contentUrl: string[]
    contentSize: number
    digest: {
        [key: string]: string
    }
    "@context": string
}